feat(ext-traits): extract new extension traits crate
