+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "cloud/core/emails",
     "cloud/core/traits",
     "cloud/email",
+    "cloud/ext-traits",
     "cloud/geo-ip",
     "cloud/proto",
     "crates/aac",
