+++ MODULE.bazel.lock
@@ -246,7 +246,7 @@
   "moduleExtensions": {
     "//vendor:cargo_ext.bzl%bindeps": {
       "general": {
-        "bzlTransitiveDigest": "iUH+t22/wPslow4xYeahkTc+DcIotqj5SFZNghuCr58=",
+        "bzlTransitiveDigest": "L2UCVIosgdkbf3kcMwiye5Zx+ECye8ertiqZ9lcHUJk=",
         "usagesDigest": "0vyNdgUB6Py8EFEYeVm1fABbu9FAMVvr905hmQzVjfU=",
         "recordedFileInputs": {},
         "recordedDirentsInputs": {},
@@ -10041,7 +10041,7 @@
     },
     "//vendor:cargo_ext.bzl%cargo_vendor": {
       "general": {
-        "bzlTransitiveDigest": "iUH+t22/wPslow4xYeahkTc+DcIotqj5SFZNghuCr58=",
+        "bzlTransitiveDigest": "L2UCVIosgdkbf3kcMwiye5Zx+ECye8ertiqZ9lcHUJk=",
         "usagesDigest": "mtR+K4/VVAmhgoN1cG7MXxv6mAD1vfxRZ8SSM66EtxE=",
         "recordedFileInputs": {},
         "recordedDirentsInputs": {},
