+++ cloud/core/cedar/src/models/organization.rs
@@ -1,7 +1,7 @@
 use core_db_types::models::{
     Organization, OrganizationInvitation, OrganizationMember, Policy, Project, Role, ServiceAccount, ServiceAccountToken,
 };
-use core_traits::OptionExt;
+use ext_traits::OptionExt;
 
 use crate::macros::{cedar_entity, cedar_entity_id};
 use crate::{CedarIdentifiable, EntityTypeName, JsonEntityUid, entity_type_name};
