+++ cloud/core/src/totp.rs
@@ -1,8 +1,8 @@
 use core_db_types::models::{MfaTotpCredential, UserId};
 use core_db_types::schema::mfa_totp_credentials;
-use core_traits::{DisplayExt, ResultExt};
 use diesel::{ExpressionMethods, QueryDsl, SelectableHelper};
 use diesel_async::RunQueryDsl;
+use ext_traits::{DisplayExt, ResultExt};
 use tonic_types::{ErrorDetails, StatusExt};
 use totp_rs::{Algorithm, TOTP, TotpUrlError};
