+++ cloud/core/traits/src/lib.rs
@@ -21,9 +21,6 @@ mod http;
 mod redis;
 mod webauthn;
 
-mod std_ext;
-pub use std_ext::*;
-
 pub trait Global:
     ConfigInterface
     + DatabaseInterface
