+++ cloud/ext-traits/BUILD.bazel
@@ -1,0 +1,8 @@
+load("//misc/utils/rust:manifest.bzl", "cargo_toml")
+load("//misc/utils/rust:package.bzl", "scuffle_package")
+
+cargo_toml()
+
+scuffle_package(
+    crate_name = "scufflecloud-ext",
+)
