+++ cloud/ext-traits/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "scufflecloud-ext-traits"
+version = "0.1.0"
+authors = ["Scuffle <opensource@scuffle.cloud>"]
+edition = "2024"
+license = "AGPL-3.0"
+publish = false
+readme = "README.md"
+repository = "https://github.com/scufflecloud/scuffle"
+description = "Extension traits for Scuffle Cloud"
+
+[dependencies]
+http = "1.3.1"
+tonic = "0.14.2"
+tonic-types = "0.14.2"
+tracing = "0.1.41"
+
+[package.metadata.sync-readme.badges]
+docs-rs = false
+crates-io = false
+license = true
+codecov = true
