+++ cloud/ext-traits/LICENSE.AGPL-3.0
@@ -1,0 +1,1 @@
+../../LICENSE.AGPL-3.0
