+++ cloud/ext-traits/README.md
@@ -1,0 +1,15 @@
+<!-- dprint-ignore-file -->
+<!-- sync-readme title [[ -->
+# scufflecloud-ext-traits
+<!-- sync-readme ]] -->
+
+<!-- sync-readme badge [[ -->
+![License: AGPL-3.0](https://img.shields.io/badge/license-AGPL--3.0-purple.svg?style=flat-square)
+[![Codecov](https://img.shields.io/codecov/c/github/scufflecloud/scuffle.svg?label=codecov&logo=codecov&style=flat-square)](https://app.codecov.io/gh/scufflecloud/scuffle)
+<!-- sync-readme ]] -->
+
+---
+
+<!-- sync-readme rustdoc [[ -->
+Extension traits.
+<!-- sync-readme ]] -->
