+++ cloud/ext-traits/src/lib.rs
@@ -1,0 +1,13 @@
+//! Extension traits.
+#![cfg_attr(coverage_nightly, feature(coverage_attribute))]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+// #![deny(missing_docs)]
+#![deny(unsafe_code)]
+#![deny(unreachable_pub)]
+#![deny(clippy::mod_module_files)]
+
+mod http_ext;
+mod std_ext;
+
+pub use http_ext::*;
+pub use std_ext::*;
